Renamed files to follow our coding style.
