+++ src/graphics/gpu_particles.cpp
@@ -15,7 +15,7 @@
 //  along with this program; if not, write to the Free Software
 //  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
-#include "graphics/gpuparticles.hpp"
+#include "graphics/gpu_particles.hpp"
 
 #include "config/user_config.hpp"
 #include "graphics/glwrap.hpp"
