+++ src/graphics/render.cpp
@@ -20,25 +20,25 @@
 
 #include "config/user_config.hpp"
 #include "graphics/callbacks.hpp"
-#include "graphics/shared_gpu_objects.hpp"
-#include "central_settings.hpp"
+#include "graphics/central_settings.hpp"
 #include "graphics/glwrap.hpp"
 #include "graphics/graphics_restrictions.hpp"
 #include "graphics/lod_node.hpp"
 #include "graphics/post_processing.hpp"
 #include "graphics/referee.hpp"
 #include "graphics/rtts.hpp"
-#include "graphics/screenquad.hpp"
+#include "graphics/screen_quad.hpp"
 #include "graphics/shaders.hpp"
-#include "graphics/shadow_matrixes.hpp"
-#include "graphics/stkmeshscenenode.hpp"
+#include "graphics/shadow_matrices.hpp"
+#include "graphics/shared_gpu_objects.hpp"
+#include "graphics/stk_mesh_scene_node.hpp"
+#include "graphics/stk_scene_manager.hpp"
 #include "items/item_manager.hpp"
+#include "items/powerup_manager.hpp"
 #include "modes/world.hpp"
 #include "physics/physics.hpp"
 #include "tracks/track.hpp"
 #include "utils/profiler.hpp"
-#include "stkscenemanager.hpp"
-#include "items/powerup_manager.hpp"
 
 #define MAX2(a, b) ((a) > (b) ? (a) : (b))
 #define MIN2(a, b) ((a) > (b) ? (b) : (a))
