+++ src/graphics/render_geometry.cpp
@@ -24,8 +24,8 @@
 #include "graphics/post_processing.hpp"
 #include "graphics/rtts.hpp"
 #include "graphics/shaders.hpp"
-#include "graphics/shadow_matrixes.hpp"
-#include "graphics/stkscenemanager.hpp"
+#include "graphics/shadow_matrices.hpp"
+#include "graphics/stk_scene_manager.hpp"
 #include "modes/world.hpp"
 #include "utils/log.hpp"
 #include "utils/profiler.hpp"
