+++ src/graphics/screen_quad.cpp
@@ -15,7 +15,7 @@
 //  along with this program; if not, write to the Free Software
 //  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
-#include "graphics/screenquad.hpp"
+#include "graphics/screen_quad.hpp"
 
 // Just the static parts to our screenquad
