+++ src/graphics/shaders.cpp
@@ -96,12 +96,13 @@
 
 #include "graphics/callbacks.hpp"
 #include "graphics/central_settings.hpp"
+#include "graphics/glwrap.hpp"
+#include "graphics/gpu_particles.hpp"
 #include "graphics/irr_driver.hpp"
-#include "graphics/gpuparticles.hpp"
 #include "graphics/shared_gpu_objects.hpp"
 #include "io/file_manager.hpp"
 #include "utils/log.hpp"
-#include "graphics/glwrap.hpp"
+
 #include <assert.h>
 #include <IGPUProgrammingServices.h>
