+++ src/graphics/stk_billboard.cpp
@@ -15,7 +15,7 @@
 //  along with this program; if not, write to the Free Software
 //  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
-#include "graphics/stkbillboard.hpp"
+#include "graphics/stk_billboard.hpp"
 
 #include "graphics/glwrap.hpp"
 #include "graphics/irr_driver.hpp"
