+++ src/graphics/stk_mesh.hpp
