+++ src/graphics/stk_scene_manager.cpp
@@ -15,17 +15,17 @@
 //  along with this program; if not, write to the Free Software
 //  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
-#include "graphics/stkscenemanager.hpp"
+#include "graphics/stk_scene_manager.hpp"
 
 #include "graphics/callbacks.hpp"
 #include "graphics/central_settings.hpp"
 #include "graphics/glwrap.hpp"
 #include "graphics/irr_driver.hpp"
 #include "graphics/lod_node.hpp"
-#include "graphics/shadow_matrixes.hpp"
-#include "graphics/stkanimatedmesh.hpp"
-#include "graphics/stkmesh.hpp"
-#include "graphics/stkmeshscenenode.hpp"
+#include "graphics/shadow_matrices.hpp"
+#include "graphics/stk_animated_mesh.hpp"
+#include "graphics/stk_mesh.hpp"
+#include "graphics/stk_mesh_scene_node.hpp"
 #include "modes/world.hpp"
 #include "tracks/track.hpp"
 #include "utils/cpp2011.hpp"
